Use AND when querying on multiple parameters at once
This involved introducing a null object for index results where the
index was not relevant, and I discovered and tightened up how we decide
whether an index is not relevant - just because the key function doesn't
throw an error doesn't mean the key isn't relevant.

I think there might still be a little leak here. When using a many-many
key, I think the key could stil return an object, but have an array
containing `[undefined]` at the leaf node. Needs more thought / testing.